api-service/pyproject.toml
[tool.poetry]
name = "api-service"
version = "0.1.0"
description = ""
authors = []
packages = [{from='src', include='api_service'}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.27.0"
pydantic-settings = "^2.2.1"
pydantic = "^2.7.0"
requests = "^2.31.0"
keycloak-client = "^0.15.4"
pytest = "^8.1.1"
fastapi-amis-admin = "^0.7.2"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.29"
sqladmin = {extras = ["full"], version = "^0.16.1"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
